az/requirements.txt
fastapi
uvicorn
azure-identity
azure-mgmt-resource
openai
python-dotenv